Consolidate ProjectExt
