+++ embrace-gradle-plugin/src/main/java/io/embrace/android/gradle/plugin/buildreporter/BuildTelemetryCollector.kt
@@ -4,9 +4,9 @@ import io.embrace.android.gradle.plugin.agp.AgpWrapper
 import io.embrace.android.gradle.plugin.config.PluginBehavior
 import io.embrace.android.gradle.plugin.gradle.GradleVersion
 import io.embrace.android.gradle.plugin.gradle.GradleVersion.Companion.isAtLeast
-import io.embrace.android.gradle.plugin.gradle.getProperty
 import io.embrace.android.gradle.plugin.instrumentation.config.model.VariantConfig
 import io.embrace.android.gradle.plugin.system.JavaSystemWrapper
+import io.embrace.android.gradle.plugin.util.getProperty
 import io.embrace.embrace_gradle_plugin.BuildConfig
 import org.gradle.api.Project
 import org.gradle.api.provider.ListProperty
@@ -93,8 +93,8 @@ class BuildTelemetryCollector {
     private fun Project.isParallelExecutionEnabled() =
         this.gradle.startParameter.isParallelProjectExecutionEnabled
 
-    private fun Project.getJvmArgs() = getProperty(GRADLE_JVM_ARGS).orNull ?: ""
-    private fun Project.getEdmVersion() = getProperty(EMBRACE_UNITY_EDM_VERSION).orNull ?: ""
+    private fun Project.getJvmArgs() = getProperty(GRADLE_JVM_ARGS) ?: ""
+    private fun Project.getEdmVersion() = getProperty(EMBRACE_UNITY_EDM_VERSION) ?: ""
 }
 
 private const val SYS_PROP_JRE_VERSION = "java.runtime.version"
