+++ embrace-gradle-plugin/src/main/java/io/embrace/android/gradle/plugin/util/ProjectExt.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.gradle.plugin.config
+package io.embrace.android.gradle.plugin.util
 
 import org.gradle.api.Project
